storage2: move MVCCComparer to config
